Don't try to (re) auth for non 401 error codes
The loginproxy only sends a WWW-Authenticate header on 401 errors. But if we try
to reauth on 404, we fail to do so, because we don't get the correct header. We
also only reach this code path if self._auth is None, which is the case if we
have a pre-authenticated cookie.