Create ClientSession inside async function api_request
Previously we were storing a `ClientSession` as a class attribute of
`Osc`. However as of https://github.com/aio-libs/aiohttp/pull/8512 that breaks,
as `aiohttp.ClientSession` requires a running asyncio event loop, which is not
guaranteed to exist. Instead we recreate a session inside `api_request` which is
an async function and is guaranteed to have a running event loop.