Include the ClientSession in the main Osc object
This prevents the returned response object from api_request using, reading using
a session that has already terminated. The session object used to be created in
a with block inside api_request, which kills it, once the function
returns. Unfortunately, that means that you cannot read larger chunks than the
default, as anything not pre-loaded tries to read from a closed session and dies.